coverage: Always process coverage, even if tests fail (#40167)
First, for projects that have both JS and PHP tests or multiple suites
of tests, run all the tests for coverage even if one fails. Mostly this
means using the `concurrently` package, since that's a very easy way to
run both and collect the exit statuses.

Then, in the workflow itself, add `always()` to the `if` on the step for
processing coverage reports so it runs even if the tests failed to
process whatever data is available.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/11842463094

Upstream-Ref: Automattic/jetpack@bbb05643e0a864672d7fddebb460551b4a8ec552